db: Switch to jackc/tern for migrations
Tern uses transactions for each migration, so a failed migration can
just be resolved by pushing a new version. Migrations with
`golang-migrate` might result in a broken database which needs to be
manually resolved.

This doesn't remove golang-migrate yet, as it runs the db-tests against
a database migrated with golang-migrate and tern, and against a database
migrated with just tern.

Fixes #197